refactor: services -> components 디렉토리명 변경
