+++ src/application/use-cases/gathering/gathering-creation.use-case.ts
@@ -3,11 +3,11 @@ import { BadRequestException, Injectable, Logger } from '@nestjs/common';
 import { APP_NAME } from 'src/common/constant';
 import { GatheringParticipationEntity } from 'src/domain/entities/gathering/gathering-participation.entity';
 import { GatheringEntity } from 'src/domain/entities/gathering/gathering.entity';
-import { GatheringInvitationsWriter } from 'src/domain/services/gathering/gathering-invitations-writer';
-import { GatheringsWriter } from 'src/domain/services/gathering/gatherings-writer';
-import { GroupsReader } from 'src/domain/services/group/groups-reader';
-import { NotificationsService } from 'src/domain/services/notification/notifications.service';
-import { UsersService } from 'src/domain/services/user/users.service';
+import { GatheringInvitationsWriter } from 'src/domain/components/gathering/gathering-invitations-writer';
+import { GatheringsWriter } from 'src/domain/components/gathering/gatherings-writer';
+import { GroupsReader } from 'src/domain/components/group/groups-reader';
+import { NotificationsService } from 'src/domain/components/notification/notifications.service';
+import { UsersService } from 'src/domain/components/user/users.service';
 import { GatheringPrototype } from 'src/domain/types/gathering.types';
 
 @Injectable()
