+++ src/application/use-cases/group/group-creation.use-case.ts
@@ -1,8 +1,8 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { APP_NAME } from 'src/common/constant';
-import { GroupsWriter } from 'src/domain/services/group/groups-writer';
-import { NotificationsService } from 'src/domain/services/notification/notifications.service';
-import { UsersService } from 'src/domain/services/user/users.service';
+import { GroupsWriter } from 'src/domain/components/group/groups-writer';
+import { NotificationsService } from 'src/domain/components/notification/notifications.service';
+import { UsersService } from 'src/domain/components/user/users.service';
 import { GroupPrototype } from 'src/domain/types/group.types';
 
 @Injectable()
