+++ src/domain/components/friend/friends-reader.ts
