+++ src/domain/components/friend/friends-writer.ts
