+++ src/domain/components/gathering/gathering-invitations-writer.ts
