+++ src/domain/components/gathering/gatherings-writer.ts
