+++ src/modules/feed/feeds.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { BlockedFeedsRepository } from 'src/domain/interface/feed/blocked-feeds.repository';
 import { FeedsRepository } from 'src/domain/interface/feed/feeds.repository';
 import { FriendFeedVisibilitiesRepository } from 'src/domain/interface/feed/friend-feed-visibilities.repository';
-import { BlockedFeedsService } from 'src/domain/services/feed/blocked-feeds.service';
-import { FeedsReader } from 'src/domain/services/feed/feeds-reader';
-import { FeedsWriter } from 'src/domain/services/feed/feeds-writer';
+import { BlockedFeedsService } from 'src/domain/components/feed/blocked-feeds.service';
+import { FeedsReader } from 'src/domain/components/feed/feeds-reader';
+import { FeedsWriter } from 'src/domain/components/feed/feeds-writer';
 import { BlockedFeedsPrismaRepository } from 'src/infrastructure/repositories/feed/blocked-feeds-prisma.repository';
 import { FeedsPrismaRepository } from 'src/infrastructure/repositories/feed/feeds-prisma.repository';
 import { FriendFeedVisibilitiesPrismaRepository } from 'src/infrastructure/repositories/feed/friend-feed-visibilities-prisma.repository';
