+++ src/modules/friend/friends.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { FriendAcceptanceUseCase } from 'src/application/use-cases/friend/friend-acceptance.use-case';
 import { FriendRequestUseCase } from 'src/application/use-cases/friend/friend-request.use-case';
 import { FriendsRepository } from 'src/domain/interface/friend/friends.repository';
-import { FriendsWriter } from 'src/domain/services/friend/friends-writer';
-import { FriendsReader } from 'src/domain/services/friend/friends-reader';
+import { FriendsWriter } from 'src/domain/components/friend/friends-writer';
+import { FriendsReader } from 'src/domain/components/friend/friends-reader';
 import { FriendsPrismaRepository } from 'src/infrastructure/repositories/friend/friends.prisma.repository';
 import { GatheringParticipationModules } from 'src/modules/gathering/gathering-participation.module';
 import { NotificationsModule } from 'src/modules/notification/notifications.module';
