+++ src/modules/gathering/gatherings.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { GatheringCreationUseCase } from 'src/application/use-cases/gathering/gathering-creation.use-case';
 import { GatheringInvitationAcceptanceUseCase } from 'src/application/use-cases/gathering/gathering-invitation-acceptance.use-case';
 import { GatheringsRepository } from 'src/domain/interface/gathering/gatherings.repository';
-import { GatheringInvitationsReader } from 'src/domain/services/gathering/gathering-invitations-reader';
-import { GatheringsReader } from 'src/domain/services/gathering/gatherings-reader';
-import { GatheringsWriter } from 'src/domain/services/gathering/gatherings-writer';
+import { GatheringInvitationsReader } from 'src/domain/components/gathering/gathering-invitations-reader';
+import { GatheringsReader } from 'src/domain/components/gathering/gatherings-reader';
+import { GatheringsWriter } from 'src/domain/components/gathering/gatherings-writer';
 import { GatheringsPrismaRepository } from 'src/infrastructure/repositories/gathering/gatherings-prisma.repository';
 import { FriendsModule } from 'src/modules/friend/friends.module';
 import { GatheringParticipationModules } from 'src/modules/gathering/gathering-participation.module';
