+++ src/modules/group/groups.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { GroupCreationUseCase } from 'src/application/use-cases/group/group-creation.use-case';
 import { GroupParticipationsRepository } from 'src/domain/interface/group/group-participations.repository';
 import { GroupsRepository } from 'src/domain/interface/group/groups.repository';
-import { GroupsWriter } from 'src/domain/services/group/groups-writer';
-import { GroupsReader } from 'src/domain/services/group/groups-reader';
+import { GroupsWriter } from 'src/domain/components/group/groups-writer';
+import { GroupsReader } from 'src/domain/components/group/groups-reader';
 import { GroupParticipationsPrismaRepository } from 'src/infrastructure/repositories/group/group-participations-prisma.repository';
 import { GroupsPrismaRepository } from 'src/infrastructure/repositories/group/groups-prisma.repository';
 import { FriendsModule } from 'src/modules/friend/friends.module';
