+++ src/modules/report/reports.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ReportsRepository } from 'src/domain/interface/report/reports.repository';
-import { FeedCommentReportsWriteService } from 'src/domain/services/report/feed-comment-reports-write.service';
-import { FeedReportsWriteService } from 'src/domain/services/report/feed-reports-write.service';
-import { FriendReportsWriteSerivce } from 'src/domain/services/report/friend-reports-write.service';
-import { GroupReportsWriteService } from 'src/domain/services/report/group-reports.write.service';
+import { FeedCommentReportsWriteService } from 'src/domain/components/report/feed-comment-reports-write.service';
+import { FeedReportsWriteService } from 'src/domain/components/report/feed-reports-write.service';
+import { FriendReportsWriteSerivce } from 'src/domain/components/report/friend-reports-write.service';
+import { GroupReportsWriteService } from 'src/domain/components/report/group-reports.write.service';
+import { ReportsWriter } from 'src/domain/components/report/reports-writer';
 import { ReportsPrismaRepository } from 'src/infrastructure/repositories/report/reports-prisma.repository';
 import { FeedsModule } from 'src/modules/feed/feeds.module';
 import { FriendsModule } from 'src/modules/friend/friends.module';
@@ -20,6 +21,7 @@ import { ReportsController } from 'src/presentation/controllers/report/reports.c
   ],
   controllers: [ReportsController],
   providers: [
+    ReportsWriter,
     FriendReportsWriteSerivce,
     GroupReportsWriteService,
     FeedReportsWriteService,
