+++ src/modules/user/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UsersRepository } from 'src/domain/interface/users.repository';
-import { UsersService } from 'src/domain/services/user/users.service';
+import { UsersService } from 'src/domain/components/user/users.service';
 import { UsersPrismaRepository } from 'src/infrastructure/repositories/users-prisma.repository';
 import { UsersController } from 'src/presentation/controllers/user/users.controller';
 @Module({
