+++ src/presentation/controllers/feed/feeds.controller.ts
@@ -26,9 +26,9 @@ import { IMAGE_BASE_URL } from 'src/common/constant';
 import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 import { AuthGuard } from 'src/common/guards/auth.guard';
 import { CreateFeedImageMulterOptions } from 'src/configs/multer-s3/multer-options';
-import { BlockedFeedsService } from 'src/domain/services/feed/blocked-feeds.service';
-import { FeedsReader } from 'src/domain/services/feed/feeds-reader';
-import { FeedsWriter } from 'src/domain/services/feed/feeds-writer';
+import { BlockedFeedsService } from 'src/domain/components/feed/blocked-feeds.service';
+import { FeedsReader } from 'src/domain/components/feed/feeds-reader';
+import { FeedsWriter } from 'src/domain/components/feed/feeds-writer';
 import { feedConverter } from 'src/presentation/converters/feed/feed.converters';
 import { BlockedFeedListRequest } from 'src/presentation/dto/feed/request/blocked-feed-list.request';
 import { CreateFriendFeedRequest } from 'src/presentation/dto/feed/request/create-friend-feed.request';
