+++ src/presentation/controllers/gathering/gatherings.controller.ts
@@ -29,10 +29,10 @@ import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 import { ApiFileOperation } from 'src/common/decorators/swagger';
 import { AuthGuard } from 'src/common/guards/auth.guard';
 import { CreateGatheringInvitationImageMulterOptions } from 'src/configs/multer-s3/multer-options';
-import { GatheringInvitationsReader } from 'src/domain/services/gathering/gathering-invitations-reader';
-import { GatheringInvitationsWriter } from 'src/domain/services/gathering/gathering-invitations-writer';
-import { GatheringsReader } from 'src/domain/services/gathering/gatherings-reader';
-import { GatheringsWriter } from 'src/domain/services/gathering/gatherings-writer';
+import { GatheringInvitationsReader } from 'src/domain/components/gathering/gathering-invitations-reader';
+import { GatheringInvitationsWriter } from 'src/domain/components/gathering/gathering-invitations-writer';
+import { GatheringsReader } from 'src/domain/components/gathering/gatherings-reader';
+import { GatheringsWriter } from 'src/domain/components/gathering/gatherings-writer';
 import { gatheringInvitationConverter } from 'src/presentation/converters/gathering/gathering-invitation.converters';
 import { gatheringConverter } from 'src/presentation/converters/gathering/gathering.converters';
 import { FileRequest, UploadImageResponse } from 'src/presentation/dto';
